+++ ffi/internal/ffi_macro.nim
@@ -2371,7 +2371,8 @@ when defined(ffiGenBindings):
     of "c":
       generateCBindings(
         genProcs, ffiTypeRegistry, libName, outDir, srcRel, ffiEventRegistry,
-        ffiConstRegistry, currentHeaderBanner,
+        ffiConstRegistry, currentHeaderBanner, ffiReverseRegistry,
+        ffiReverseEventRegistry,
       )
     of "cddl":
       generateCddlBindings(genProcs, ffiTypeRegistry, libName, outDir, srcRel)